new: Updated base color palette
Added new color palette to base.json.

## Brand colors
Brand colors are indicated in the description i.e. "brand anchor: brand color name". Brand colors serve as soft anchor points for chroma and lightness that influence neighboring tones and tints.

## Color families
Color families are built as needed according to brand color hue, and filling in gaps for colors required by product teams, i.e. Red, Green (status notifications), Indigo (visualizations), etc. Purple, Azure, Amber, Blue, Teal, Orange, Coral, Magenta are palettes with 1 or more brand color included.

## Tonal scale
Colors use an expanded tonal scale that go from 25, 50, 100 ... 900, 950, 975, with colors going from lightest to darkest. Each step should be used for a specific purpose - see table in Figma for mapping.

## Accessibility
Colors are designed in okLCH, a perceptually uniform color space that more accurately predicts contrast relationships. The following pairs are guaranteed to meet text contrast requirements

- 0, 600
- 100, 600
- 200, 700
- 300, 800
- 400, 900
- 500, 1000

And for non-text contrast

- 0, 500
- 25, 500 (Input borders)
- 200, 600
- 300, 700
- 400, 800
- 500, 900
- 600, 1000

This is true regardless of color family (red, blue, green, slate, etc).

## Neutrals

Two neutrals are included - slate and neutral. Slate is a brand tinted neutral that uses Workday's brand blue to create. We do this by decreasing the strength of the curve to peak at around 0.034 chroma at 500.

Neutral is a pure greyscale that has 0 chroma or hue, going from pure white to black. This is meant for prose content - text that is not meant to be branded. We also use this as the base for things like overlays, translucent surfaces, etc.

## Color names
Colors are named to be readable for both humans and the browser. Color names are composed of two parts: the color family and the color grade. Branded names (like cinnamon/blueberry/ink etc) require a level of cultural context to understand which may not be shared with all teams. Moving towards simple color names make it easier for everyone to be on the same page when discussing color, as well as making it easier to integrate for acquisition teams who may be coming in with their own design system.

## Chroma and hue

Chroma is the intensity of the color. Hue is the degree that represents a color on the color wheel (0 to 360). Chroma is a value that ranges from 0 to infinity - maxChroma of a color depends on the hue and lightness of the color, as well as the color profile of the device. okLCH is gamut-independent, meaning we can define any color in the visible spectrum. But want to make sure to define colors that can be rendered by devices - so ensuring that color defined are clipped to the sRGB or p3 color space.

Colors are defined to peak in chroma at the 500-600 step to create a natural sense of progression through the palette, with chroma decreasing at the ends. Accent colors should use the 600 color to ensure visible foregrounds while remaining highly saturated.

Amber is a yellow - these colors peak in chroma much earlier on. For these colors, 400 is more appropriate for use as accent backgrounds, and a darker foreground instead of white. These colors also shift in hue as we progress to darker colors. This is done to increase the maxChroma at each step.