new: base tokens for spacing
Adding two base tokens for new space tokens to be built from.

`base.fontSize` is the default font size (16px/1rem). `base.baseline` is for the baseline grid to align to. We can probably mark `base.unit` as deprecated in the future (if we want), as `baseline` is essentially it's replacement. We are just making the relationship to the 8pt grid clearer for space/size tokens.