new: transparent tokens
Adding new transparency tokens for Hover and Active states of Inverse Secondary Buttons.

## Changes

We changed our bg.transparent color token to use black instead of white, and added a "default" suffix to the token to avoid JSON nesting bug. Changing it to black and fully transparent let us add two other variants for the hover and active states, strong and stronger.

We kept `bg.overlay` and `bg.transcluscent` the same - adding variants of those didn't make sense as we are not creating variants of modal with lighter scrims overlays, and these are not colors that are only slightly opaque.